// SDKs/twilio.com/twilio_verify_v2/1.37.3/typescript/src/sdk/models/operations/listratelimit.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";
import * as shared from "../shared";

export const ListRateLimitServerList = [
	"https://verify.twilio.com",
] as const;


export class ListRateLimitPathParams extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=ServiceSid" })
  serviceSid: string;
}


export class ListRateLimitQueryParams extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "queryParam, style=form;explode=true;name=PageSize" })
  pageSize?: number;
}


export class ListRateLimitSecurity extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "security, scheme=true;type=http;subtype=basic" })
  accountSidAuthToken: shared.SchemeAccountSidAuthToken;
}


export class ListRateLimitListRateLimitResponseMeta extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=first_page_url" })
  firstPageUrl?: string;

  @SpeakeasyMetadata({ data: "json, name=key" })
  key?: string;

  @SpeakeasyMetadata({ data: "json, name=next_page_url" })
  nextPageUrl?: string;

  @SpeakeasyMetadata({ data: "json, name=page" })
  page?: number;

  @SpeakeasyMetadata({ data: "json, name=page_size" })
  pageSize?: number;

  @SpeakeasyMetadata({ data: "json, name=previous_page_url" })
  previousPageUrl?: string;

  @SpeakeasyMetadata({ data: "json, name=url" })
  url?: string;
}


export class ListRateLimitListRateLimitResponse extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=meta" })
  meta?: ListRateLimitListRateLimitResponseMeta;

  @SpeakeasyMetadata({ data: "json, name=rate_limits", elemType: shared.VerifyV2ServiceRateLimit })
  rateLimits?: shared.VerifyV2ServiceRateLimit[];
}


export class ListRateLimitRequest extends SpeakeasyBase {
  @SpeakeasyMetadata()
  serverUrl?: string;

  @SpeakeasyMetadata()
  pathParams: ListRateLimitPathParams;

  @SpeakeasyMetadata()
  queryParams: ListRateLimitQueryParams;

  @SpeakeasyMetadata()
  security: ListRateLimitSecurity;
}


export class ListRateLimitResponse extends SpeakeasyBase {
  @SpeakeasyMetadata()
  contentType: string;

  @SpeakeasyMetadata()
  listRateLimitResponse?: ListRateLimitListRateLimitResponse;

  @SpeakeasyMetadata()
  statusCode: number;
}
